Extract the common slug_to_person behaviour into a module
