+++ lib/membership_csv/people.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 require 'csv'
-require_relative 'person'
+require_relative '../people_slug_to_person'
 
 module MembershipCSV
   class People
@@ -8,6 +8,8 @@ module MembershipCSV
       @csv_filename = csv_filename
       @person_factory = person_factory
     end
+
+    include PeopleSlugToPerson
 
     def find_all
       all_people.sort_by(&:name)
@@ -47,24 +49,6 @@ module MembershipCSV
 
     def id_to_person
       @id_to_person ||= all_people.map { |person| [person.id, person] }.to_h
-    end
-
-    def slug_to_person
-      @slug_to_person ||= all_people.each_with_object({}) { |person, memo| ensure_unique_slug(person, memo) }
-    end
-
-    def ensure_unique_slug(person, memo)
-      raise_if_no_slug(person)
-      raise_if_repeated_slug(person, memo)
-      memo[person.slug] = person
-    end
-
-    def raise_if_no_slug(person)
-      raise "No slug for #{person.name}" unless person.slug
-    end
-
-    def raise_if_repeated_slug(person, memo)
-      raise "Slug #{person.slug} repeated for #{memo[person.slug].name} and #{person.name}" if memo[person.slug]
     end
 
     def mapit_id_to_person
